backend/pyproject.toml
[project]
name = "multiagent-backend"
version = "0.1.0"
description = "FastAPI backend for multi-agent workshop (prototype-first)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "python-dotenv>=1.0",
  "pydantic>=2.7",
  "httpx>=0.27",
  "duckduckgo-search>=6.2",
  "openai>=1.51",
  "anthropic>=0.40",
  "google-generativeai>=0.7",
  "ollama>=0.3",
  # Frameworks (kept minimal, can be commented out if needed)
  "langchain>=0.3",
  "langgraph>=0.2",
  "smolagents>=0.1.0",
  "python-multipart>=0.0.9",
  "langchain-openai>=0.1.23",
]

[tool.uv]
# No special config; uv sync inside backend dir

